+++ CHANGELOG.md
@@ -13,7 +13,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security -->
 
-## [1.1.1] - 2020-05-07
+## [1.3.0] - 2020-05-12
+
+### Added
+
+- `ign` command.
+
+### Changed
+
+- `tpdms` and `tpll` are now unified.
+
+### Removed
+
+- `tpdms` command.
+
+
+## [1.2.0] - 2020-05-07
 
 ### Added
